Crate zwave_volatile_attr_poll_manager

This module reads yaml configuration file that contains attribute types and its polling interval values in [seconds] as follow: Keys for polling intervals: … polling_interval_zwave -> represents the interval for Z-Wave node … polling_interval_zwave_v1 -> represents the interval for Z-Wave Plus v1 node … polling_interval_zwave_v2 -> represents the interval for Z-Wave Plus v2 node

Example registering polling interval for classic Z-Wave and Zwave Plus v1 device classes:

  - attribute_type: ATTRIBUTE_COMMAND_CLASS_BINARY_SWITCH_STATE
    polling_interval_zwave: 100
    polling_interval_zwave_v1: 100

Example registering polling interval for all device classes:

   - attribute_type: ATTRIBUTE_COMMAND_CLASS_BINARY_SWITCH_STATE
     polling_interval_zwave: 100
     polling_interval_zwave_v1: 100
     polling_interval_zwave_v2: 30

[PollRegister] will watch for included nodes in the attribute store. When a new node is included or gets updated. It will schedule a poll interval for that node. Given the user configured a poll interval for that node matching the zwave plus info version number.

Functions

zwave_poll_manager_init

Initialize the attribute poll manager return sl_status_t SL_STATUS_OK on success